requirements.txt
dj-database-url==0.4.1
Django==1.9.7
django-crispy-forms==1.6.1
django-registration-redux==1.4
gunicorn==19.6.0
olefile==0.43
Pillow==4.0.0
psycopg2==2.6.1
sorl-thumbnail==12.3
whitenoise==2.0.6
